Move executor into separate module
